requirements.txt
# Core app
streamlit==1.37.0
PyMuPDF==1.24.10
python-docx==1.1.2
pandas==2.2.2
numpy==1.26.4
python-dateutil==2.9.0.post0
tqdm==4.66.4
pytz==2024.1
python-dotenv==1.0.1
yagmail==0.15.293

# Vector DB
chromadb==0.5.23

# LLM + embeddings
openai>=1.30.0,<2
sentence-transformers==2.6.0
torch==2.3.1           

# Crew
crewai==0.157.0

# DOCX fallbacks
mammoth==1.6.0
docx2txt==0.8

# API
fastapi==0.111.0
uvicorn==0.30.0

# DB
SQLAlchemy>=2.0.31,<2.1
psycopg2-binary==2.9.9

# testing
pytest==8.2.2
scikit-learn==1.4.2
